lr_gym/requirements_sb3.txt
stable-baselines3==1.4.0
torch==1.10.1
torchvision==0.11.2
rospkg==1.3.0
catkin_pkg==0.4.24
tensorboard==2.3.0
defusedxml==0.6.0
gym==0.19.0
mpi4py==3.0.3
tqdm==4.59.0
pybullet==2.8.1
nptyping==1.3.0
numpy-quaternion==2020.5.19.15.27.24
opencv-python==4.5.5.62
empy==3.3.4
numba==0.53.0
PyVirtualDisplay==2.1
pynvml==11.0.0
playsound
quaternionic==0.4.1
pyyaml==6.0
